added the api for deletion of event and imported only required functions from different modules
