+++ club/models.py
@@ -3,7 +3,6 @@ from django.conf import settings
 from django.contrib.auth.models import User
 from django.db.models.signals import post_save
 from django.dispatch import receiver
-from users.models import *
 import datetime
 
 class Club(models.Model):
