+++ events/views.py
@@ -1,27 +1,20 @@
 from django.shortcuts import render,get_object_or_404
 from rest_framework import viewsets,generics,permissions
-from .models import *
+from events.models import Event
+from club.models import Club
 from messages import *
 from keys import *
 from django.contrib.auth.models import User
 from django.contrib.auth import login,authenticate
 from rest_framework.authtoken.models import Token
 from django.http import HttpResponse,JsonResponse
-from django.views.decorators.csrf import csrf_exempt
 from rest_framework.views import APIView
-from rest_framework.response import Response
 from rest_framework.authentication import SessionAuthentication, BasicAuthentication,TokenAuthentication
 from rest_framework.decorators import api_view, permission_classes
 from rest_framework.permissions import AllowAny
 from django.shortcuts import redirect
-from django.contrib.sites.shortcuts import get_current_site
-from django.utils.encoding import force_text,force_bytes
-from django.utils.http import urlsafe_base64_encode,urlsafe_base64_decode
-from django.template.loader import render_to_string
 from django.core.mail import send_mail,EmailMessage
-from django.urls import reverse_lazy
-import random
-import json
+from random import randint
 
 class CreateEvent(APIView):
     """  
@@ -38,6 +31,8 @@ class CreateEvent(APIView):
         event_pic=request.FILES.get(ep_)
         event_start_date=request.POST.get(esd_)
         event_end_date=request.POST.get(eed_)
+        if event_end_date is None:
+            event_end_date=event_start_date
         club_obj=Club.objects.get(id=club)
         print(club_obj.id)
         event=Event.objects.create(club=club_obj,title=title,short_desc=short_desc,description=description,venue=venue,event_pic=event_pic,event_start_date=event_start_date,event_end_date=event_end_date)
@@ -54,7 +49,7 @@ class EditEvent(APIView):
     permission_classes=(permissions.IsAuthenticated,)
     authentication_classes=(TokenAuthentication,)
     def post(self,request,*args,**kwargs):
-        event_id=request.GET.get(event_id_)
+        event_id=request.POST.get(event_id_)
         club=request.POST.get(ci)
         title=request.POST.get(title_)
         short_desc=request.POST.get(sd_)
@@ -78,6 +73,21 @@ class EditEvent(APIView):
             'success':True,
             'message':events.ecsuccess
             })
+
+class EventDelete(APIView):
+    """  
+    This class is for deleting a existing event.
+    """
+    permission_classes=(permissions.IsAuthenticated,)
+    authentication_classes=(TokenAuthentication,)
+    def get(self,request,*args,**kwargs):
+        event_id=request.GET.get(event_id_)
+        event_obj=Event.objects.get(id=event_id)
+        event_obj.delete()
+        return JsonResponse({
+            'success':True,
+            'message':events.delete
+        })
 
 class EventList(APIView):
     """  
@@ -129,3 +139,4 @@ class EventDetails(APIView):
             'event_start_date':event_obj.event_start_date,
             'event_end_date':event_obj.event_end_date,
         })
+
