+++ news/urls.py
@@ -1,0 +1,6 @@
+from django.urls import path,include
+from . import views
+from django.conf.urls import url
+urlpatterns=[
+    
+]
