+++ users/admin.py
@@ -1,5 +1,8 @@
 from django.contrib import admin
-from .models import *
+from users.models import UserProfile,WelcomeModel
+from club.models import Club,ClubMember,ClubFollower
+from news.models import News
+from events.models import Event
 
 class ClubsAdmin(admin.ModelAdmin):
     list_display=('name','tagline','created_date','modified_date')
