+++ users/urls.py
@@ -8,6 +8,5 @@ urlpatterns=[
     path('login/',view=views.Login.as_view()),
     path('forgot_password/',view=views.ForgotPassword.as_view()),
     path('reset_password/',view=views.ResetPassword.as_view()),
-    path('postnews',view=views.PostNews.as_view()),
     
 ]
