+++ users/views.py
@@ -1,28 +1,19 @@
 from django.shortcuts import render,get_object_or_404
 from rest_framework import viewsets,generics,permissions
-from club.models import *
-from .models import *
+from users.models import *
 from messages import *
 from keys import *
 from django.contrib.auth.models import User
 from django.contrib.auth import login,authenticate
 from rest_framework.authtoken.models import Token
 from django.http import HttpResponse,JsonResponse
-from django.views.decorators.csrf import csrf_exempt
 from rest_framework.views import APIView
-from rest_framework.response import Response
 from rest_framework.authentication import SessionAuthentication, BasicAuthentication,TokenAuthentication
 from rest_framework.decorators import api_view, permission_classes
 from rest_framework.permissions import AllowAny
 from django.shortcuts import redirect
-from django.contrib.sites.shortcuts import get_current_site
-from django.utils.encoding import force_text,force_bytes
-from django.utils.http import urlsafe_base64_encode,urlsafe_base64_decode
-from django.template.loader import render_to_string
 from django.core.mail import send_mail,EmailMessage
-from django.urls import reverse_lazy
-import random
-import json
+from random import randint
 
 class Welcome(APIView):
     """ 
@@ -59,7 +50,7 @@ class Register(APIView):
                     'message':users.rfailure
                 })
             else:
-                otp=random.randint(999,999999)
+                otp=randint(999,999999)
                 user.set_password(password)
                 user.save()
                 y=UserProfile.objects.get(user=user)
@@ -78,7 +69,7 @@ class Register(APIView):
         user=User.objects.create_user(username=email,password=password,email=email)
         user.is_active=False
         user.save()
-        otp=random.randint(999,999999)
+        otp=randint(999,999999)
         y=UserProfile.objects.get(user=user)
         y.name=name
         y.otp=str(otp)
@@ -164,7 +155,7 @@ class ForgotPassword(APIView):
                 'success':False,
                 'message':users.fpfailure
             })
-        otp=random.randint(999,999999)
+        otp=randint(999,999999)
         user=User.objects.get(username=email)
         up=UserProfile.objects.get(user=user)
         up.otp=otp
@@ -210,18 +201,3 @@ class ResetPassword(APIView):
                 #'access_token':None
             })
 
-
-class PostNews(APIView):
-    permission_classes=(permissions.IsAuthenticated,)
-    authentication_classes=(TokenAuthentication,)
-    def post(self,request,*args,**kwargs):
-        club=request.POST.get(ci)
-        title=request.POST.get(title_)
-        short_desc=request.POST.get(sd_)
-        description=request.POST.get(desc_)
-        news_pic=request.FILES.get(np_)
-        #user=request.user
-        club_obj=Club.objects.get(id=club)
-        news=News.objects.create(club=club_obj,title=title,short_desc=short_desc,description=description,news_pic=news_pic)
-        news.save()
-        return JsonResponse({'message':'news has been posted'})
